Fix compiling on Windows

+++ src/animation.c
@@ -23,7 +23,7 @@
 #include "ai.h"
 #include "network.h"
 #include "audio.h"
-#include "time.h"
+#include "zgtime.h"
 #include "globals.h"
 #include "window.h"
 
@@ -418,7 +418,7 @@ static void moveWeapon(Weapon *weapon, double timeDelta)
 		}
 		
 		weapon->z = INITIAL_WEAPON_Z + fabsf(cosf(weapon->timeFiring * 16.0f)) * 1.5f;
-		weapon->timeFiring += timeDelta;
+		weapon->timeFiring += (float)timeDelta;
 	}
 }
 
@@ -508,7 +508,7 @@ static void animateTilesAndPlayerRecovery(double timeDelta, ZGWindow *window, Ch
 		
 		/* Then when animation_timer reaches BEGIN_DESTROYING_TILES, decrement the colored tile's z value by TILE_FALLING_SPEED, disable the tile's state, enable the tile's recovery_timer, and give the tile a proper recovery time delay */
 		
-		if (player->needTileLoc && player->animation_timer >= (BEGIN_DESTROYING_TILES - player->weap->compensation) && player->player_loc != -1)
+		if (player->needTileLoc && player->animation_timer >= (BEGIN_DESTROYING_TILES - (double)player->weap->compensation) && player->player_loc != -1)
 		{
 			player->destroyedTileIndex = player->player_loc;
 			player->needTileLoc = false;
